Working Linkbar with health=100
All damage should be of type float for now on, the linkbar needs to be at health=100 to scale correctly as of now, but it's fixed

+++ Assets/Scripts/Character/Player/Kuze/Kuze.cs
@@ -67,7 +67,7 @@ public class Kuze : APlayer
         // _moveFastAbility.Toggle();
     }
 
-    public override void Damage(int damage)
+    public override void Damage(float damage)
     {
         linkBarUI.RemoveLink(damage);
         base.Damage(damage);
